[PATCH] Add file build-support/.nosearch
This makes sure that the bundled copy of `elisp-lint.el' does not end
up on the `load-path'.

Bundling third-party libraries is seldom a good idea because they
tend not to get updated.  And that might mean that a user who also has
the stand-alone version installed end up loading the outdated version
instead.  This commit at least makes that less likely.

[Emacsmirror] This commit is being rebased onto upstream.